Make the Arithmetic example more newbie-friendly
This commit adds a build.xml so that it is easy to build by ant and updates the
example so that it can parse arithmetic expressions passed in as an
argument. The overall intention is to bring this example to a state where it can
be used in a CongoCC tutorial.